refactor: KakaoSignUpResult 모델 SocialSignUpResult로 변경 및 userId 저장 로직 추가
- KakaoSignUpResult 모델명을 SocialSignUpResult로 변경하여 카카오 외 다른 소셜 로그인에도 사용 가능하도록 범용성 확대
- 소셜 로그인 성공 시 응답 헤더에서 userId를 추출하여 DataStore에 저장하는 로직 추가

Related files:
- data/src/main/kotlin/com/sooum/data/repository/SocialLoginRepositoryImpl.kt
- domain/src/main/kotlin/com/sooum/domain/repository/SocialLoginRepository.kt
- domain/src/main/kotlin/com/sooum/domain/usecase/kakao/KakaoSignUpUseCase.kt
- data/src/main/kotlin/com/sooum/data/network/socialLogin/SocialLoginApi.kt
- domain/src/main/kotlin/com/sooum/domain/usecase/kakao/NaverSignUpUseCase.kt
- domain/src/main/kotlin/com/sooum/domain/model/auth/SocialSignUpResult.kt (기존 KakaoSignUpResult.kt)
