+++ data/src/main/kotlin/com/sooum/data/repository/AuthRepositoryImpl.kt
@@ -5,25 +5,16 @@ import com.sooum.data.datastore.AppManageDataStore
 import com.sooum.data.network.auth.AuthApi
 import com.sooum.data.network.auth.request.EmailVerifyRequest
 import com.sooum.data.network.auth.request.LoginRequest
-import com.sooum.data.network.auth.request.NameOnlyRequest
 import com.sooum.data.network.auth.request.SignUpRequest
 import com.sooum.data.network.safeFlow
 import com.sooum.domain.model.ApiResult
 import com.sooum.domain.model.CheckEmail
-import com.sooum.domain.model.EmailVerifyResult
-import com.sooum.domain.model.KakaoSignUpResult
 import com.sooum.domain.model.LoginResult
-import com.sooum.domain.model.PostProfileResult
 import com.sooum.domain.model.PostRefreshTokenResult
 import com.sooum.domain.model.SignUpResult
 import com.sooum.domain.repository.AuthRepository
 import jakarta.inject.Inject
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flowOf
-import okhttp3.MediaType.Companion.toMediaTypeOrNull
-import okhttp3.MultipartBody
-import okhttp3.RequestBody.Companion.asRequestBody
-import java.io.File
 
 class AuthRepositoryImpl @Inject constructor(
     private val authApi: AuthApi,
