+++ domain/src/main/kotlin/com/sooum/domain/model/auth/SocialSignUpResult.kt
@@ -1,9 +1,9 @@
-package com.sooum.domain.model
+package com.sooum.domain.model.auth
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class KakaoSignUpResult (
+data class SocialSignUpResult(
     val userId: Int,
     val signUp: Boolean,
     val profileImage: String?
