+++ domain/src/main/kotlin/com/sooum/domain/repository/AuthRepository.kt
@@ -1,14 +1,10 @@
 package com.sooum.domain.repository
 
 import com.sooum.domain.model.ApiResult
-import com.sooum.domain.model.EmailVerifyResult
-import com.sooum.domain.model.KakaoSignUpResult
 import com.sooum.domain.model.LoginResult
-import com.sooum.domain.model.PostProfileResult
 import com.sooum.domain.model.PostRefreshTokenResult
 import com.sooum.domain.model.SignUpResult
 import kotlinx.coroutines.flow.Flow
-import java.io.File
 
 interface AuthRepository {
