+++ domain/src/main/kotlin/com/sooum/domain/usecase/kakao/KakaoSignUpUseCase.kt
@@ -1,7 +1,7 @@
 package com.sooum.domain.usecase.kakao
 
 import com.sooum.domain.model.ApiResult
-import com.sooum.domain.model.KakaoSignUpResult
+import com.sooum.domain.model.auth.SocialSignUpResult
 import com.sooum.domain.repository.SocialLoginRepository
 import jakarta.inject.Inject
 import kotlinx.coroutines.flow.Flow
@@ -12,7 +12,7 @@ class KakaoSignUpUseCase @Inject constructor(
     suspend operator fun invoke(
         accessToken: String,
         refreshToken: String
-    ): Flow<ApiResult<KakaoSignUpResult>> {
+    ): Flow<ApiResult<SocialSignUpResult>> {
         return kakaoRepository.kakaoLogin(accessToken, refreshToken)
     }
 }
